+++ Tdf/TdfInteger.h
@@ -6,19 +6,19 @@
 class TdfInteger : public Tdf
 {
 private:
-	QWORD m_value = 0;
+	DWORD m_value = 0;
 
 public:
 	TdfInteger();
-	TdfInteger(DWORD Label, TdfTypes Type, QWORD Value);
+	TdfInteger(DWORD Label, TdfTypes Type, DWORD Value);
 	~TdfInteger();
 
 	static TdfInteger* fromMemory(void* buffer, DWORD * size = nullptr);
 
 	virtual DWORD toMemory(void* buffer, DWORD size);
 
-	QWORD getValue() { return m_value; }
-	void setValue(QWORD Value) { m_value = Value; }
+	DWORD getValue() { return m_value; }
+	void setValue(DWORD Value) { m_value = Value; }
 };
 
 #endif //TDFINTEGER_H
