+++ Tdf/TdfIntegerList.h
@@ -8,19 +8,19 @@
 class TdfIntegerList : public Tdf
 {
 private:
-	std::vector<QWORD> m_values;
+	std::vector<DWORD> m_values;
 
 public:
 	TdfIntegerList();
-	TdfIntegerList(DWORD Label, TdfTypes Type, std::vector<QWORD> Values);
+	TdfIntegerList(DWORD Label, TdfTypes Type, std::vector<DWORD> Values);
 	~TdfIntegerList();
 
 	static TdfIntegerList* fromMemory(void* buffer, DWORD * size = nullptr);
 
 	virtual DWORD toMemory(void* buffer, DWORD size);
 
-	std::vector<QWORD> getValues() { return m_values; }
-	void setValues(std::vector<QWORD> Values) { m_values = Values; }
+	std::vector<DWORD> getValues() { return m_values; }
+	void setValues(std::vector<DWORD> Values) { m_values = Values; }
 };
 
 #endif //TDFINTEGERLIST_H
